Implement parallel routing for drops (#203)
* implement parallel routing for drops

* loading ui cleanup
