+++ src/app/(project)/projects/[project]/drops/(drop)/[drop]/edit/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import EditDrop from '../../../EditDrop';
+
+interface EditDropProps {
+  params: { drop: string; project: string };
+}
+
+export default function EditDropPage({ params: { drop, project } }: EditDropProps) {
+  return (
+    <div className="w-max mx-auto pt-20">
+      <EditDrop drop={drop} project={project} />
+    </div>
+  );
+}
