+++ src/app/(project)/projects/[project]/drops/layout.tsx
@@ -1,0 +1,14 @@
+export default function DropsLayout({
+  children,
+  modal,
+}: {
+  children: React.ReactNode;
+  modal: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+      {modal}
+    </>
+  );
+}
